Remove unnecessary double casting

+++ src/main/java/fonda/scheduler/scheduler/data/TaskData.java
@@ -126,7 +126,7 @@ public class TaskData implements Comparable<TaskData> {
     private void calc() {
         if ( nodeDataTuples.isEmpty() ) value = Double.MIN_VALUE;
         if ( size == 0 ) value = 1;
-        else value = nodeDataTuples.get(0).getSizeInBytes() / (double) size;
+        else value = nodeDataTuples.get(0).getSizeInBytes() / size;
     }
 
     @Override
